% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_vcf_vep.R
\name{annotate_vcf_vep}
\alias{annotate_vcf_vep}
\title{Function for variant annotation from vcf file}
\usage{
annotate_vcf_vep(
  path,
  pval = 5e-08,
  dir_cache =
    "/cloud-data/its-cmo-darwin-magellan-workspaces-folders/WS_genetic_ts/PMCB_Genetics/data/VEP/110",
  cache_version = 110,
  nThread
)
}
\arguments{
\item{path}{Specify the full path starting with /cloud-home/ to the vcf file, including the file name.}

\item{pval}{p value threshold to filter. Default = 5e-8}

\item{dir_cache}{Specify the cache directory to use. Required. Default: "/cloud-data/its-cmo-darwin-magellan-workspaces-folders/WS_genetic_ts/PMCB_Genetics/data/VEP/110"}

\item{cache_version}{a numeric. Use a different cache version than the assumed default (the VEP version). Required. Default: 110}

\item{nThread}{Number of threads to use for parallel processes}
}
\value{
a data.frame for variant annotation. ensemblVEP will return 82 columns from VEP.
}
\description{
A wrapper for variant annotation by VEP via bioconductor ensemblVEP
}
\references{
\url{https://grch37.ensembl.org/info/docs/tools/vep/script/vep_options.html}

\url{https://bioconductor.org/packages/release/bioc/vignettes/ensemblVEP/inst/doc/ensemblVEP.pdf}
}
