% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_vcf.R
\name{mr_vcf}
\alias{mr_vcf}
\title{Function for Mendelian randomization using vcf file}
\usage{
mr_vcf(
  exposure_vcf,
  outcome_vcf,
  pval = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  ldref,
  plink_bin = genetics.binaRies::get_plink_binary(),
  exclude_HLA = TRUE,
  mr_method = subset(TwoSampleMR::mr_method_list(), use_by_default)$obj[c(1, 2, 4)],
  nThread
)
}
\arguments{
\item{exposure_vcf}{Specify the full path starting with /cloud-home/ to the vcf file, including the file name, for exposure. Normally a molecular qtl file}

\item{outcome_vcf}{Specify the full path starting with /cloud-home/ to the vcf file, including the file name, for outcome. Normally a gwas file}

\item{pval}{p value threshold to extract the instruments from exposure vcf file Default = 5e-8}

\item{clump_r2}{Clumping r2 threshold. Default: 0.001}

\item{clump_kb}{Clumping kb window. Default: 10000. (very strict)}

\item{ldref}{LD reference data. Normally Europeans from the 1000 genomes reference panel. Default:"/cloud-home/U1006974/NeuroID/data/1kg/EUR"}

\item{plink_bin}{Specify path to plink binary. Default = "genetics.binaRies::get_plink_binary()"}

\item{exclude_HLA}{Exclude SNPs in HLA region or not. Default: TRUE. GRCh37; Location: chr6:28,477,797-33,448,354;}

\item{mr_method}{methods to use in mr analysis. See [mr_method_list()] for details. Default: subset(mr_method_list(), use_by_default)$obj[c(1,2,4)].
it will contain "mr_wald_ratio", "mr_egger_regression" and "mr_ivw"}

\item{nThread}{Number of threads to use for parallel processes}
}
\value{
a data.frame for mr results
}
\description{
Extract instruments from exposure vcf file by a pvalue threshold and do clumping
and then extract them from outcome vcf file, format both datasets into TwosampleMR
run Mendelian randomization using default methods, "mr_wald_ratio", "mr_egger_regression" and "mr_ivw"
}
\references{
\url{https://mrcieu.github.io/gwasglue/articles/mr.html}
}
