% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstatsColHeaders}
\alias{sumstatsColHeaders}
\title{Summary Statistics Column Headers}
\format{
dataframe with 2 columns
}
\usage{
data("sumstatsColHeaders")
}
\description{
List of uncorrected column headers often found in GWAS Summary
Statistics column headers. Note the effect allele will always be the A2
allele, this is the approach done for
VCF(https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7805039/).
(https://github.com/neurogenomics/MungeSumstats/blob/HEAD/R/data.R)
This is enforced with the column header corrections here and also the check allele flipping
test.
}
\keyword{datasets}
