% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcftibble_to_TwoSampleMR.R
\name{vcftibble_to_TwoSampleMR}
\alias{vcftibble_to_TwoSampleMR}
\title{Function for creating exposure or outcome data format for TwoSampleMR from a tibble/data.frame}
\usage{
vcftibble_to_TwoSampleMR(df, type)
}
\arguments{
\item{df}{a data.frame or a tibble}

\item{type}{"exposure" or "outcome"}
}
\value{
a data.frame for format of TwoSampleMR
}
\description{
creating exposure or outcome data format for TwoSampleMR from a tibble
which is created by vcf_to_granges(vcf) %>% dplyr::as_tibble()
}
